move settings to its own direcotry in screens
