+++ src/screens/SettingsSheet/components/DevSection.tsx
@@ -14,13 +14,13 @@ import {
 // @ts-ignore
 import Restart from 'react-native-restart';
 import { useDispatch } from 'react-redux';
-import { defaultConfig } from '../../config/experimental';
-import useAppVersion from '../../hooks/useAppVersion';
-import { settingsUpdateNetwork } from '../../redux/settings';
+import { defaultConfig } from '@rainbow-me/config/experimental';
+import useAppVersion from '../../../hooks/useAppVersion';
+import { settingsUpdateNetwork } from '@rainbow-me/redux/settings';
 import NetworkSection from './NetworkSection';
-import Menu from './components/Menu';
-import MenuContainer from './components/MenuContainer';
-import MenuItem from './components/MenuItem';
+import Menu from './Menu';
+import MenuContainer from './MenuContainer';
+import MenuItem from './MenuItem';
 import { WrappedAlert as Alert } from '@/helpers/alert';
 import { deleteAllBackups } from '@/handlers/cloudBackup';
 import { getProviderForNetwork, web3SetHttpProvider } from '@/handlers/web3';
@@ -35,7 +35,7 @@ import {
   useUpdateAssetOnchainBalance,
   useWallets,
 } from '@/hooks';
-import { ImgixImage } from '@/components/images';
+import { ImgixImage } from '@rainbow-me/images';
 import { wipeKeychain } from '@/model/keychain';
 import { clearAllStorages } from '@/model/mmkv';
 import { Navigation } from '@/navigation';
@@ -45,16 +45,16 @@ import { clearImageMetadataCache } from '@/redux/imageMetadata';
 import store from '@/redux/store';
 import { walletsUpdate } from '@/redux/wallets';
 import { ETH_ADDRESS } from '@/references';
-import Routes from '@/navigation/routesNames';
+import Routes from '@rainbow-me/routes';
 import { ethereumUtils } from '@/utils';
-import logger from '@/utils/logger';
+import logger from 'logger';
 import {
   removeNotificationSettingsForWallet,
   useAllNotificationSettingsFromStorage,
   addDefaultNotificationGroupSettings,
 } from '@/notifications/settings';
 import { IS_DEV } from '@/env';
-import { SettingsLoadingIndicator } from '@/components/settings-menu/SettingsLoadingIndicator';
+import { SettingsLoadingIndicator } from '@/screens/SettingsSheet/components/SettingsLoadingIndicator';
 
 const DevSection = () => {
   const { navigate } = useNavigation();
