+++ src/screens/SettingsSheet/components/MenuItem.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Source } from 'react-native-fast-image';
-import { ButtonPressAnimation } from '../../animations';
-import CheckmarkCircledIcon from '../../icons/svg/CheckmarkCircledIcon';
-import WarningIcon from '../../icons/svg/WarningIcon';
+import { ButtonPressAnimation } from '../../../components/animations';
+import CheckmarkCircledIcon from '../../../components/icons/svg/CheckmarkCircledIcon';
+import WarningIcon from '../../../components/icons/svg/WarningIcon';
 import Chevron from '@/assets/chevronUpDown.png';
 import Caret from '@/assets/family-dropdown-arrow.png';
 import { Box, Inline, Stack, Text } from '@/design-system';
-import { ImgixImage } from '@/components/images';
+import { ImgixImage } from '@rainbow-me/images';
 import { useTheme } from '@/theme';
 
 interface ImageIconProps {
