+++ src/screens/SettingsSheet/components/PrivacySection.tsx
@@ -1,12 +1,12 @@
 import React, { useCallback, useReducer } from 'react';
 import { Switch } from 'react-native-gesture-handler';
-import Menu from './components/Menu';
-import MenuContainer from './components/MenuContainer';
-import MenuItem from './components/MenuItem';
+import Menu from './Menu';
+import MenuContainer from './MenuContainer';
+import MenuItem from './MenuItem';
 import useExperimentalFlag, { PROFILES } from '@/config/experimentalHooks';
 import { useAccountProfile, useShowcaseTokens, useWebData } from '@/hooks';
 import { useNavigation } from '@/navigation';
-import Routes from '@/navigation/routesNames';
+import Routes from '@rainbow-me/routes';
 import { device } from '@/storage';
 import * as i18n from '@/languages';
 import { analyticsV2 } from '@/analytics';
