+++ src/screens/SettingsSheet/constants.ts
@@ -1,0 +1,1 @@
+export const CUSTOM_MARGIN_TOP_ANDROID = 8;
