+++ src/screens/SettingsSheet/settingsCardStyleInterpolator.ts
@@ -1,0 +1,36 @@
+import { StackCardInterpolationProps } from '@react-navigation/stack';
+import { Animated } from 'react-native';
+
+export function settingsCardStyleInterpolator({
+  current,
+  next,
+  inverted,
+  layouts: { screen },
+}: StackCardInterpolationProps) {
+  const translateFocused = Animated.multiply(
+    current.progress.interpolate({
+      inputRange: [0, 1],
+      outputRange: [screen.width, 0],
+    }),
+    inverted
+  );
+  const translateUnfocused = next
+    ? Animated.multiply(
+        next.progress.interpolate({
+          inputRange: [0, 1],
+          outputRange: [0, -screen.width],
+        }),
+        inverted
+      )
+    : 0;
+
+  return {
+    cardStyle: {
+      transform: [
+        {
+          translateX: Animated.add(translateFocused, translateUnfocused),
+        },
+      ],
+    },
+  };
+}
